Final Commit
Ready for Review. Changed the file names to their proper names. Added in the script to the index. Added in a function that is called when the window is loaded. That function adds click events to all of the order buttons. Added a function that adds items to the cart. If the item already exists in the cart then it adds one to the order count. Commented the javascript file.